+++ android/app/src/main/java/com/asiafreshman2018/MainActivity.java
@@ -1,6 +1,8 @@
 package com.asiafreshman2018;
 
+import android.os.Bundle;
 import com.facebook.react.ReactActivity;
+import org.devio.rn.splashscreen.SplashScreen;
 
 public class MainActivity extends ReactActivity {
 
@@ -9,7 +11,9 @@ public class MainActivity extends ReactActivity {
      * This is used to schedule rendering of the component.
      */
     @Override
-    protected String getMainComponentName() {
-        return "AsiaFreshman2018";
+    protected String getMainComponentName() { return "AsiaFreshman2018"; }
+    protected void onCreate(Bundle savedInstanceState) {
+        SplashScreen.show(this);  // here
+        super.onCreate(savedInstanceState);
     }
 }