+++ src/pages/FriendlyCampus/CampusSecurity.js
@@ -8,7 +8,7 @@ class CampusSecurity extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '校園安全',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });