+++ src/pages/HappySchool/SchoolTuitionReduction.js
@@ -9,7 +9,7 @@ class SchoolTuitionReduction extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '就學優待減免',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });