+++ src/pages/LearningDevelopment/CareerDevelopment.js
@@ -9,7 +9,7 @@ class CareerDevelopment extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '生涯發展與就業輔導',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });