+++ src/pages/LearningDevelopment/Course.js
@@ -36,7 +36,7 @@ class Course extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '選課',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });