+++ src/pages/LearningDevelopment/Course/Category.js
@@ -9,7 +9,7 @@ class Category extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '本校課程分類表',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });