+++ src/pages/LearningDevelopment/Course/RequiredCourses.js
@@ -9,7 +9,7 @@ class RequiredCourses extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '校定必修課程注意事項',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });