+++ src/pages/LearningDevelopment/Course/Timetable.js
@@ -8,7 +8,7 @@ class Timetable extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '新生時程說明及注意事項',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });