+++ src/pages/LearningDevelopment/SoftwareResources.js
@@ -8,7 +8,7 @@ class SoftwareResources extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '電子郵件 & 軟體資源',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });