+++ src/pages/MainItem/CampusLife.js
@@ -26,7 +26,7 @@ class CampusLife extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '校園生活',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.navigate({routeName: 'Homepage', actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });